Mostly changing defaults and adding doco
git-svn-id: https://svn.atnf.csiro.au/askapsdp/Src/trunk@3339 1db04d4c-0ff0-4329-ae0b-98c546476c6a

+++ Code/Components/Synthesis/synthesis/current/deconvolution/DeconvolverControl.tcc
@@ -133,7 +133,7 @@ namespace askap {
       this->setTargetObjectiveFunction(parset.getFloat("targetobjective", 0.0));
       this->setFractionalThreshold(parset.getFloat("fractionalthreshold", 0.0));
       this->setLambda(parset.getFloat("lambda", 0.0001));
-      this->setPSFWidth(parset.getFloat("psfwidth", 0));
+      this->setPSFWidth(parset.getInt32("psfwidth", 0));
     }
 
   } // namespace synthesis
